# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'Blaaize Wallet API'
  description: 'This is the API documentation for Blaaize Wallet.'
  version: 1.0.0
paths:
...
  /api/user/profile:
    get:
      tags:
        - User
      summary: 'Get User Profile with Wallets'
      operationId: b875a04f8f8d99781211d13821b130f4
      responses:
        200:
          description: 'User profile retrieved successfully'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    properties:
                      user:
                        properties:
                          id:
                            type: integer
                            example: 1
                          firstname:
                            type: string
                            example: John
                          lastname:
                            type: string
                            example: Doe
                          email:
                            type: string
                            format: email
                            example: johndoe@example.com
                        type: object
                      wallets:
                        type: array
                        items:
                          properties:
                            id:
                              type: integer
                              example: 1
                            currency:
                              type: string
                              example: USD
                            balance:
                              type: number
                              example: 100.5
                          type: object
                    type: object
                  message:
                    type: string
                    example: 'User profile retrieved successfully'
                type: object
        401:
          description: Unauthenticated
        403:
          description: Forbidden
      security:
        -
          sanctum: []
  /api/wallet/create:
    post:
      tags:
        - Wallet
      summary: 'Create a new wallet'
      description: 'Creates a wallet for the authenticated user with the specified currency.'
      operationId: createWallet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - currency
              properties:
                currency:
                  type: string
                  example: USD
              type: object
      responses:
        201:
          description: 'Wallet created successfully'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Wallet created successfully.'
                  wallet_id:
                    type: integer
                    example: 1234
                type: object
        400:
          description: 'Invalid currency or wallet already exists'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Invalid currency or wallet already exists.'
                type: object
        401:
          description: Unauthorized
        500:
          description: 'Internal Server Error'
      security:
        -
          sanctum: []
  /api/wallet/list:
    get:
      tags:
        - Wallet
      summary: 'List all wallets'
      description: 'Fetches a paginated list of wallets for the authenticated user.'
      operationId: listWallets
      parameters:
        -
          name: perPage
          in: query
          description: 'Number of wallets per page'
          required: false
          schema:
            type: integer
            example: 10
      responses:
        200:
          description: 'Wallets retrieved successfully'
          content:
            application/json:
              schema:
                properties:
                  current_page:
                    type: integer
                    example: 1
                  total:
                    type: integer
                    example: 50
                  per_page:
                    type: integer
                    example: 10
                  last_page:
                    type: integer
                    example: 5
                  data:
                    type: array
                    items:
                      properties:
                        id:
                          type: integer
                          example: 123
                        currency:
                          type: string
                          example: USD
                        balance:
                          type: number
                          format: float
                          example: 1000.5
                        created_at:
                          type: string
                          format: date-time
                          example: '2023-01-01T12:00:00Z'
                      type: object
                type: object
        401:
          description: Unauthorized
        500:
          description: 'Internal Server Error'
      security:
        -
          sanctum: []
  '/api/wallet/list/{walletId}':
    get:
      tags:
        - Wallet
      summary: 'View a single wallet'
      description: 'Fetches details of a specific wallet belonging to the authenticated user.'
      operationId: viewWallet
      parameters:
        -
          name: walletId
          in: path
          description: 'ID of the wallet to retrieve'
          required: true
          schema:
            type: integer
            example: 1
      responses:
        200:
          description: 'Wallet retrieved successfully'
          content:
            application/json:
              schema:
                properties:
                  id:
                    type: integer
                    example: 1
                  currency:
                    type: string
                    example: USD
                  balance:
                    type: number
                    format: float
                    example: 1000.5
                  created_at:
                    type: string
                    format: date-time
                    example: '2023-01-01T12:00:00Z'
                  updated_at:
                    type: string
                    format: date-time
                    example: '2023-01-01T12:00:00Z'
                type: object
        404:
          description: 'Wallet not found'
        401:
          description: Unauthorized
        500:
          description: 'Internal Server Error'
      security:
        -
          sanctum: []
components:
  securitySchemes:
    sanctum:
      type: http
      bearerFormat: JWT
      scheme: bearer
...
security:
  -
    sanctum:
      type: apiKey
      description: 'Enter token in format (<token>)'
      name: Authorization
      in: header